requirements.txt
Django==1.9.7
PyJWT==1.4.1
django-webpack-loader==0.3.0
djangorestframework==3.4.0
wsgiref==0.1.2
